tehtävä4/requirements.txt
psycopg2-binary
contextlib
